set_timezone() add openrc, close https://github.com/Chrysostomus/manjaro-architect/issues/41

+++ lib/util-base.sh
@@ -77,7 +77,11 @@ set_timezone() {
     DIALOG " $_ConfBseTimeHC " --yesno "$_TimeZQ ${ZONE}/${SUBZONE}?" 0 0
 
     if [[ $? -eq 0 ]]; then
-        arch_chroot "ln -s /usr/share/zoneinfo/${ZONE}/${SUBZONE} /etc/localtime" 2>$ERR
+        if [[ -e /tmp/.openrc ]]; then
+            arch_chroot "ln -s /usr/share/zoneinfo/posix/${ZONE}/${SUBZONE} /etc/localtime" 2>$ERR
+        else
+            arch_chroot "ln -s /usr/share/zoneinfo/${ZONE}/${SUBZONE} /etc/localtime" 2>$ERR
+        fi
         check_for_error
     else
         config_base_menu
